+++ src/include/parser.hpp
@@ -16,6 +16,8 @@ class Parser {
 
   Number readValue();
 
+  Number readNonNegativeValue();
+
   std::string readIdentifier();
 
   Operand readOperand();